requirements.txt
six==1.12.0
websocket-client==0.54.0